TODO: Only write and perform tests
